chore(services): make interfaces more generic
I replaced all of the `&Path` and `Vec<String>` arguments with
`impl AsRef<Path>` and `&[impl AsRef<str>]`. The reason is that this
allows you to pass in more types at the call site. For example, the
`impl AsRef<Path>` argument allows you to pass in `&Path`, `PathBuf`,
or `&PathBuf` at the call site, which is more flexible. Similarly, the
`&[impl AsRef<str>]` allows you to pass in static strings in addition
to owned strings.

I also changed the call to process-compose to use the `output` method
instead of `spawn` followed by `wait`. The `output` method does both
of these already and simply waits for the process to terminate, which
saves you an extra `map_err` and potential error.

I also changed the `debug!` call to attach the process names as span
data, as this will still get printed in log messages and is how we
handle printing data associated with a log message in most other places.
We never fully migrated from `log::*` to `tracing::*` macros, so you
may see the old style of string formatting, but that should be avoided
for new code.