refactor(buildenv): cleanup a few names
In several places, "package" is used to refer to the `RealisedPackage`
for one of a package's outputs. I find referring to a package's packages
confusing, so use "outputs" instead in these cases.

Rename `originalPackage` -> `storePathsToInstallIds` since this is the
only thing it's used for, and also change `std::pair<std::string,
resolver::LockedPackageRaw>` -> `std::string` since the
`LockedPackageRaw` was unused.