feat: generate errors when stopping services
Errors are parsed from `process-compose` logs in order to determine the type of failure that occurred. This allows us to extract service names from logs to inform the user when they've tried to stop a service that either doesn't exist or is already stopped.

Currently `process-compose` doesn't differentiate between trying to stop a service that doesn't exist and trying to stop a service that's already stopped, but as a later enhancement we can use `process-compose process list` to disambiguate.

This also removes a duplicate tag from `services.bats` as it already has a `file_tags` directive at the top.