feat: unknown service error types (#1759)
## Proposed Changes
- Replace the client generated `MessageType` with an `untagged` type
with an `Unknown` variant.
- Add an `UnknownServiceMessage` to the `ResolutionFailure` enum
- Remove the `Option` on the `MessageLevel` field in related structs.
It's required in the API, so it should always be present.

As a follow up, I think the message handling shouldn't be tied
explicitly to _failures_ since we can pass warnings and additional
explanatory messages. But that can be a follow up to refactor that when
it's needed.

## Release Notes

- Improved resolution failure messaging.

---------

Co-authored-by: Yannik Sander <yannik@floxdev.com>