feat(services): flox services stop command (#1763)
## Proposed Changes

Adds the `flox services stop` command, which is accessible behind the
services feature flag.
- `flox service stop [NAME, ...]` will stop the named services
- `flox services stop` will stop all user-specified services

Notably this _does not_ shut down `process-compose`. `process-compose`
automatically shuts itself down if all the running services have
completed, which is inconvenient for a number of reasons, so instead we
inject a dummy process that sleeps forever. The intention is that a
later ticket that implements automatic cleanup will shut down
`process-compose`.

This also contains the plumbing for the `flox services status` command
as reading the output of `process-compose process list` is necessary for
proper error handling.

## Release Notes

N/A

---------

Co-authored-by: Zach Mitchell <zmitchell@fastmail.com>