feat: support building flake installables
When deserializing v1 manifests, use `locked-flake-attr-path` to
determine whether to treat a package as a catalog or flake package.

Flake packages don't get attribute path defaulting of `legacyPackages`,
and they don't get wrapped as a flox-nixpkgs input.

`locked-system` was renamed to `system` for consistency between catalog
and flake locked packages.