feat: support building flake installables (#1784)
When deserializing v1 manifests in `buildenv`, use
`locked-flake-attr-path` to
determine whether to treat a package as a catalog or flake package.

Flake packages don't get attribute path defaulting of `legacyPackages`,
and they don't get wrapped as a flox-nixpkgs input.

`locked-system` was renamed to `system` for consistency between catalog
and flake locked packages.

TEST_DATA_DIR was renamed to GENERATED_DATA in

https://github.com/flox/flox/commit/3338caad08b554fb3757f90ebb1589b753fe12be,
but now we need data from both
sources, so restore the usage of TEST_DATA_DIR.