chore: Remove Flox.access_tokens
Plus the corresponding `nix` structure in `flox.toml`. We don't
`deny_unknown_fields` on the config so it won't cause errors in the
unlikely event that anyone still has values set.

Adjust the config tests to use `trusted_environments` as an example of
removing absent keys.

The last consumer of this was removed in 46c49495.

I noticed this while looking at places that we might use or set HTTP
credentials in relation to Graham's issue with installing from flakes.