test(services): Redis on standard port
Test a networked service. We originally deferred this because of port
clashes when running on non-containerised CI systems (Nix builders) but:

a) it's important that we test this most common use case for services
b) workaround for ports will be added in a subsequent commit

I've chosen Redis because it's relatively small and fast to start,
unlike PostgreSQL for example. It currently creates a data dir within
the temporary project environment but that could be disabled with the
`--save "" --appendonly no` arguments.

I considered using `envs` instead of `resolve` to store the manifest
alongside the mocks but it required more setup and didn't make it much
clearer than co-locating the manifest with the tests.