fix(activate): Fix flox_portgrab args for nc
The version of netcat that we get takes the port as a positional
argument. The test was failing because it was unable to match the
expected output against the usage output:

    % /nix/store/aj3i0mqh3cfp2yr16kn034s9c6p1mzbi-libressl-3.8.2-nc/bin/nc 0.0.0.0 -l 0 -n -v
    usage: nc [-46cDdFhklNnrStUuvz] [-C certfile] [-e name] [-H hash] [-I length]
	     [-i interval] [-K keyfile] [-M ttl] [-m minttl] [-O length]
	     [-o staplefile] [-P proxy_username] [-p source_port] [-R CAfile]
	     [-s sourceaddr] [-T keyword] [-V rtable] [-W recvlimit] [-w timeout]
	     [-X proxy_protocol] [-x proxy_address[:port]] [-Z peercertfile]
	     [destination] [port]

This was a bit of a pain to debug because:

- changes to the scripts require a rebuild of the environment
- the errors can swallowed by the redirect
- debugging via `flox activate -v` is quite limited

Which, combined with `on-activate` being run multiple times, makes me
feel like we shouldn't release this helper just yet.