test(services): Redis on ephemeral port
Prevent port clashes on non-containerised CI systems (Nix builders) by
using Michael's `flox_portgrab` function.

I'm not quite comfortable with exposing this to users yet because
subsequent activations will get a different port value for the client.

We could work around it by providing some sugar to store a named value
in `$FLOX_ENV_CACHE` but it locks us in to providing a specific
interface and it would be better to revise whether replaying activation
values is feasible first.

So for now I've copied the script internally to the tests and provided
`netcat` through the environment. The only differences from the draft PR
are the heading comment and using an unqualified `nc` path.