test(services): Poll for process-compose list
The new Redis test was hitting the same kind of race condition that we
saw in 1452e922 whereby the socket is created just shortly before the
service is actually started. So it was failing with the output:

   -- command failed --
   status : 1
   output (3 lines):
     Waiting for process-compose socket
     Could not connect to Redis at 127.0.0.1:52928: Connection refused
     Shutting down process-compose
   --

Prevent this from happening by polling `process-compose process list` as
we recently discussed doing when we get to #1640.