feat: support priority and systems in flake descriptors (#1796)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->

**feat: allow flake descriptors to specify systems**

This mostly amounts to simply only generating flake installables for the
requested systems rather than simply generating installables for all of
the systems in the manifest.

This change also adds priorities to the descriptor type, but doesn't
respect them in pkgdb yet.

**chore: manual impl of InstallableLocker**

There's only one method in this trait, so implementing the trait on the
enum and manually dispatching is very easy and saves us from needing to
rely on a macro.

**feat: support priority in flake descriptors**

Applies a priority to flake packages using this order:
- Priority set in the descriptor
- `meta.priority` of the derivation
- Default priority (5, for now)

**refactor: make priority a u64**

The Nix codebase explicitly declares priority as a NixInt, which is
declared as a uint64_t. This changes the Rust code to be explicit about
matching that. In practice `usize` and `u64` are the same, but it
doesn't hurt to match the semantics set by Nix exactly.

**test: make small test smaller**

This test now takes very little time because it doesn't need nixpkgs.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->