feat: skip starting services when already running
This uses the presence of the socket file to determine when to skip
starting services. A warning message is printed to inform the user that
they requested to start services as part of the activation and we
intentionally did not do what they requested.

In order to make this easy to test I reverted the socket path function
to the previous behavior where it would defer to an explicitly set
environment variable.

It would be nice to test that `process-compose` is also not started
when we show this message, but that seems tricky to test in a
cross-platform way (e.g. no access to `/proc` on macOS).