feat: prototype manifest build
DO NOT MERGE

This patch implements a PoC of the manifest build which uses
the instructions as found in the `[build]` section to build
the package exactly as it would be built by manually invoking
the same commands within a `flox activate` shell.

- sets $out in the environment
- invokes the build commands in a subshell (using bash)
- turns the $out directory into a Nix package with all outpath
  references replaced with the real $out
- wraps all `bin/*` commands with `$FLOX_ENV/activate`

Usage: flox-build [ pkgname ]

N.B. this patch utilises the (hitherto-unused) $FLOX_TURBO flag
to invoke the provided command without an intervening shell, and
in the process I tripped across and fixed a bug in the activate
script.

fix: fixes & incorporate Tom's suggestions

Plus split out bash script to separate file.

feat: new approach, use Makefile to implement multi-stage builds

feat: substitute path references for package dependencies in builds

This patch adds the feature of rewriting the build script to substitute
all references to `${pkgname}` within a build script with their storePath
values as derived from `readline result-pkgname` prior to kicking off the
builds. With this feature we can automatically infer the dependency graph
and configure builds to depend upon each other.

fix: support for impure builds per Tom's suggestion

fix: optimize INTERMEDIATE targets, fix rewriting of dependency packages

Tested, now works. Only wrinkle is it will use a result symlink if it exists
rather than recompile unconditionally.

fix: use our versions of gnu utilities in preference to Darwin

fix: don't be quite so noisy

feat: cache build directory across builds

This feature adds a new optional buildCache output to the flox sandbox
build and returns a link to it alongside the usual result symlink. Flox
then provides this path back into subsequent builds so that files can
be copied into place to make subsequent compilations faster.

feat: various

- redesign for local and sandbox builds only with optional
  buildCache for pure repeatable sandbox builds
  - ensure that the temporary buildCache always has same content-
    addressed storePath
  - ensure that the tar image does not capture any timestamps relating
    to directory entries
- add support to `activate` script for setting $0 for wrapped scripts
- be more careful about `${install-prefix}` provided by user, ensure it
  exists and allow it to be a file
- update the result symlink target to always rebuild the requested
  targets only
- fix `activate bash <script>` invocations to use `-e` flag so that script
  failures result in a nonzero outcome
- prevent our MAKEFLAGS from being inherited by the build's invocations
  of make

fix: refactor to pass dependencies as buildInputs

I had forgotten to pass the identified inter-package dependencies as
buildInputs to the runCommand invocation. This commit does that, and
as it happens also required a reordering of processing so that the
$(_pvarname)_buildscript variable was defined before evaluating the
dependencies. This patch may look like a lot of lines changed, but
it is mostly just the order of processing that has changed.

fix: ensure that buildCache is initialized to be unique

We ran into the problem that init'ing a buildcache would essentially
trick Nix into believing that a previous initialization was valid to
be reused for a later initialization, and while we're unsure whether
that's a desired behaviour we're certain that it's an unexpected one.

This update will "seed" all newly-initialized build caches with a
.buildCache.init file containing the UTC timestamp of initialization,
which should guarantee the uniqueness of all buildCaches and avoid
this unexpected behaviour when attempting to discard the buildCache.

fix: refactor to catch changes in source

If we use stdenv to pull in the source we lose the timestamps in
the process, and we rely upon those timestamps to know when to
rebuild objects in the buildCache with respect to their source.
So we instead pass in the source as a tarball to be extracted in
the sandbox at the time of the build.

This effort also brought about the realisation that I was making
the whole stable buildCache thing harder than it needed to be.
All we need to do is ensure we use the same filename across subsequent
builds, and then let Nix do the checksumming and validation to check
when things change.

Also refactored the makefile to split out template hierarchy as
follows:

  - DEPENDS_template
  - BUILD_template
    - BUILD_local_template
    - BUILD_sandbox_template

... and then selected the choice of local or sandbox template within
the loop as we iterate over $(BUILDS). Also updated a load of comments,
hopefully makes things clearer.

feat: virtual sandbox

The Flox "virtual sandbox" warns or aborts when encountering an ELF access
from outside the closure of packages implied by $FLOX_ENV. In this regard
it can provide the same guarantees at an ELF level provided by the sandbox
itself, but at an _advisory_ level, so that developers are informed of
missing dependencies without actually breaking anything.

The virtual sandbox is enabled by setting `FLOX_VIRTUAL_SANDBOX=(warn|abort)`
in the environment, and we do this when wrapping files in the bin directory
in the course of performing a manifest build.

As with the parsing of FLOX_ENV_LIB_DIRS, it is essential that this parsing
of the closure be performant and initialized only once per invocation, so we
start by reading closure paths into a btable from $FLOX_ENV/requisites.txt.

fix: missed setting sandbox in plain build

feat: support sandbox builds on Darwin and Linux

Fixes included:
- cannot use libsandbox in pure mode .. do not need it anyway
- remove old buildCache when creating next one
- add support for FLOX_SANDBOX_ALLOW_DIRS
- revert sandbox features in ld-floxlib.so now that we have it
  in the standalone libsandbox.{so,dylib}
- sign Mac binaries after performing sed substitution
- avoid unstable gzip compression, fix makefile targets
- added placeholder makedot script