feat: block activate while starting services
This uses helper functions in the `activate` script to block while
waiting for `process-compose` to respond to an initial status request.
It also makes the following notable changes:
- Adds the `_FLOX_SERVICES_ACTIVATE_TIMEOUT` variable
- Adds the `_FLOX_SERVICES_LOG_FILE` variable
- Redirects `process-compose` logs to a file

The `_FLOX_SERVICES_ACTIVATE_TIMEOUT` variable is used to override the
default activation timeout of 1s. Note that `process-compose` itself
doesn't block on creating the socket, so if activation takes 1s then
something has gone very wrong in a way that's likely unrecoverable.
This variable is useful during testing to kill `process-compose` when
you've put it into a state such that it will *never* start properly e.g.
you've set the socket path to a path that needs elevated write
permissions.

The `_FLOX_SERVICES_LOG_FILE` variable is used to set an explicit path
for the `process-compose` logs. The default behavior is to write to a
temp file.

When activation fails the user is shown an error message that may or may
not contain the path to the log file. The reason for this is that
`process-compose` may exit before writing anything to the log file. In
that case you don't want to direct the user to read logs in a
nonexistent file.

In terms of implementation there's also some saving/restoring of shell
settings so attempting to parse an error message as JSON doesn't exit
the activation script abruptly. We also need to set `NO_COLOR` when
starting `process-compose` otherwise it will fill the log file with
escape codes.