feat(watchdog): initialize with logger
This creates a separate watchdog binary called `klaus`. The first
attempt created a `flox watchdog` subcommand, but getting a simple
file-based logger integrated with the existing `flox` logging machinery
turned out to be quite a hassle (and didn't work, even when it
compiled). It also required looking at the parsed arguments *before*
dispatching on the subcommand to configure the logger since the logger
is configured very early in the startup process.

In short, creating a separate binary just provides much more flexibility
and simplicity compared to a `flox` subcommand since `flox` is doing
much more than this application will.