feat: use getopt, consistent -c handling for all cmd modes
This PR and the discussions around the use of FLOX_TURBO highlighted
the reasoning behind our decision to invoke commands following the
`--` delimiter, and further established this as exactly equivalent
to the `-c` argument handling provided by all shells. This patch
updates the `activate` script to implement this same handling,
making it possible to invoke commands using either of the `--` or
`-c` notations in any of the `--turbo`, `--noprofile`, or normal
"cmd" modes.

This change makes it possible to successfully invoke:

  % FLOX_TURBO=1 /path/to/activate bash <file>

Also renamed FLOX_NO_PROFILES -> FLOX_NOPROFILE env var for
consistency both with `activate` and `bash` cmdline options.