feat: be smarter about detecting when to rebuild activation-scripts
This small tweak of the pkgdb/Makefile changes the logic so that the
activation-scripts package is only rebuilt when any of its assets or
its Nix expression changes, which saves about 4 seconds with each
invocation of make in the pkgdb subdirectory.

It also does a small amount of reorganization by putting all of the
makefile stanzas pertaining to `realise.o` in the same location.