fix: don't lock v0 manifests (#1812)
The goal of this commit is to avoid ever locking with pkgdb. But we
still want to support several operations on v0 environments, namely:
- `flox list` (implemented by `deserialized_lockfile`)
- `flox activate`  (implemented by `activation_path`)
- `flox containerize` (implemented by `build_container`)

Currently, `flox list` and `flox activate` will re-lock a v0 manifest.
There was a regression in `flox containerize` for an unlocked v0
manifest, and it will now error with:
```
❌ ERROR: couldn't canonicalize path ".../.flox/env/manifest.lock": No such file or directory (os error 2): No such file or directory (os error 2)
```

To support these operations without locking, add a method
`ensure_locked` for path and managed environments, and use
`ensure_locked` instead of `lock` for list, activate, and containerize.
`ensure_locked` will error for managed and remote environments if the
lockfile is missing in a generation.

If there's an unlocked v0 environment, throw an error, e.g.:
```
> flox list
❌ ERROR: Locking version 0 manifests is no longer supported.
Set 'version = 1' in the manifest.
```

`lock` is still used in all transactions that modify environments, in
`create_generation_from_local_env`, and in `push_new`, as we don't need
to support v0 manifests for any of those operations.

Since building has to be performed without re-locking,
ManagedEnvironment::build was split up into `lock`, `build`, and `link`.
An explicit call to `lock` was added where necessary.

`build` was removed from the environment trait as it was unused. `lock`
was replaced with `deserialized_lockfile`, since operations that depend
on the lockfile (i.e. `list`) shouldn't always actually lock the
environment.

If a `ManagedEnvironment` has a lockfile out of sync with the latest
generation, it needs a re-lock. `ManagedEnvironment::validate_checkout`
was modified to check for this condition, although previously an out of
sync lockfile was not treated as an out of sync local checkout.

An error is now thrown if `use_catalog == true` and an attempt is made
to lock a v0 manifest.

manifest.toml and manifest.lock files that could be generated without a
catalog client were added to a new `test_data/manually_generated`
directory. These were used in several places to replace now-deprecated
pkgdb lock calls.