fix: close fd3 in process-compose test
process-compose 1.6.1 has a race condition when started with a socket
that it cannot write to. It sometimes leaves running services behind.

Bats will hang forever if a child process doesn't exit because it's
waiting for file descriptor 3 to close. See:
https://bats-core.readthedocs.io/en/stable/writing-tests.html#file-descriptor-3-read-this-if-bats-hangs

Close FD 3 so that tests don't hang forever if the sleep process is left
behind.

$ cat service-config.yaml
processes:
  sleep:
    command: "sleep infinity"
$ ps -A | grep sleep
$ while ! ps -A | grep sleep; do process-compose up --config service-config.yaml --unix-socket /no-perms; done
24-07-26 20:50:45.882 FTL start UDS http server on /no-perms failed error="listen unix /no-perms: bind: permission denied"
...
24-07-26 20:50:47.319 FTL start UDS http server on /no-perms failed error="listen unix /no-perms: bind: permission denied"
1588157 pts/1    00:00:00 sleep