refactor: build klaus in flake, use in flox
This refactors the flake to make the package that builds derivations
from our Rust source generic over the crate being built. This allows
us to share all the setup required to configure our Nix-built Rust
applications.

The `flox-rust-sdk` and `catalog-api-v1` crates are built twice,
but the shared cargo dependencies are built once and reused.