test(services): Poll for watchdog start and finish
To prevent race conditions where we send SIGINT before the signal
handler is registered or check the PID no longer exists before the
watchdog has had a chance to respond.

I removed the assertion on the state output because the behaviour varies
between platform. MacOS appears to use it to indicate that the process
has no controlling terminal, as reproduced with:

    % sleep infinity &
    [2] 68796

    * % ps -p $! -o state=
    ?

Linux does something different though:

    dan@build-arm.floxdev.com ~ $ sleep infinity &
    [1] 3515846
    dan@build-arm.floxdev.com ~ $ ps -p $! -o state=
    S