fix: address race condition on watchdog launch
There was a race condition on launching the watchdog in which the parent
process could terminate between launching the watchdog and it installing
its signal handler and listening for termination. This means that the
watchdog would be reparented to PID 1, and would never exit (because PID
1 only exits on a shutdown of the machine).

This mitigates the race condition by always passing the activation's
PID to the watchdog and having the watchdog check whether that PID is
still running shortly after startup. If the provided PID isn't running,
then the watchdog immediately exits. There's still a race condition
between the time of *checking* that the PID is running and the time of
*deciding* to continue with watchdog execution, but that window is a
handful of instructions compared to a fork, exec, initializing the
watchdog logger, etc.