fix: refactor to pass dependencies as buildInputs
I had forgotten to pass the identified inter-package dependencies as
buildInputs to the runCommand invocation. This commit does that, and
as it happens also required a reordering of processing so that the
$(_pvarname)_buildscript variable was defined before evaluating the
dependencies. This patch may look like a lot of lines changed, but
it is mostly just the order of processing that has changed.