refactor(watchdog): Always use PID from flox
We know ahead of time what the activation PID will be when we start the
watchdog because Flox execs into the activation and keeps the same PID.
So we don't need to infer it from the watchdog.

This prevents race conditions where the activation may have exited
before the watchdog completely starts and it gets re-parented to PID1.
When this happens we exit early with an error if the PID that was passed
is no longer the parent of the watchdog.