refactor(services): Move logic for prevent remote
Previously we relied on `services_socket_path()` giving us an error for
remote environments but we now need to always pass a socket path to the
watchdog regardless of whether it does or will exist.

I was hoping to implement this on `ServicesCommands` but we don't have
access to the environment there. It's kinda ugly to throw away the `dyn
Environment` in activate but it serves the most common use case where we
need it for `services` commands.