feat(watchdog): Track activations in registry (#1837)
## Proposed Changes

**refactor(registry): Take reg_path for lock_path**

So that we don't need to construct `Flox` in `klaus` when accessing the
registry. All of the existing callers already have the registry path
when constructing the lock path.

**feat(klaus): Pass in dot_flox_hash for registry**

So that we can access the environment registry without having to
reconstruct information from `Flox` or `.flox`.

**feat(registry): Track activation PIDs**

This will be used by the watchdog `klaus` to track when the last
activation of an environment exits and we should clean up any running
services.

Stale activations are removed when updating the registry on disk to
catch watchdog processes that may have died without cleaning up after
themselves.

The new field needs to use `default` in order to read registry files
from existing Flox versions where the field isn't present.

Some assumptions in the implementation:

- environments are already registered by the time we're activating or
  starting the watchdog; an error will be returned if not
- an activation will only be registered and deregistered once;
  subsequent calls are noops
- we can't record activations for remote environments because they
  aren't currently registered in the environment registry; that's not
  currently an issue for service management because we don't support
  services for remote environments at the moment
- testing focuses on unit testing the inner parts because the file
  handling part is relatively simple, proven, and more fiddly to
  co-ordinate in a series of tests

**refactor(watchdog): Make socket mandatory**

We'll later have klaus check whether the socket exists in order to
determine whether it needs to shutdown `process-compose` to allow for
environments that start services imperatively after the watchdog has
been started.

refactor(env): Expose dot_flox_path()

Needed to construct an environment registry path hash that we can pass
to the watchdog.

**refactor: Wire activation registration into watchdog**

Now that the watchdog has been merged. Uses the newly added
`dot_flox_path()`. This appears to work for remote envs too, which I
previously thought weren't registered.

**refactor(watchdog): Always use PID from flox**

We know ahead of time what the activation PID will be when we start the
watchdog because Flox execs into the activation and keeps the same PID.
So we don't need to infer it from the watchdog.

This prevents race conditions where the activation may have exited
before the watchdog completely starts and it gets re-parented to PID1.
When this happens we exit early with an error if the PID that was passed
is no longer the parent of the watchdog.

**refactor(services): Move logic for prevent remote**

Previously we relied on `services_socket_path()` giving us an error for
remote environments but we now need to always pass a socket path to the
watchdog regardless of whether it does or will exist.

I was hoping to implement this on `ServicesCommands` but we don't have
access to the environment there. It's kinda ugly to throw away the `dyn
Environment` in activate but it serves the most common use case where we
need it for `services` commands.

## Release Notes

N/A (scaffolding for an actual feature)

---------

Co-authored-by: Zach Mitchell <zmitchell@fastmail.com>