feat: automatically shut down services
After deregistering an environment, check if the list of activations is
now empty and if so send the shutdown command to `process-compose`.

This also redirects the watchdog stdout/stderr streams to `/dev/null`
when launched by `flox` so that none of the watchdog output appears
in the shell (especially when there's an error).