fix: ensure that buildCache is initialized to be unique
We ran into the problem that init'ing a buildcache would essentially
trick Nix into believing that a previous initialization was valid to
be reused for a later initialization, and while we're unsure whether
that's a desired behaviour we're certain that it's an unexpected one.

This update will "seed" all newly-initialized build caches with a
.buildCache.init file containing the UTC timestamp of initialization,
which should guarantee the uniqueness of all buildCaches and avoid
this unexpected behaviour when attempting to discard the buildCache.