feat(services): impl Display for ProcessStates
Using the output format described in:

- https://www.notion.so/floxdev/Services-status-output-53ee184e2b5b4b7d8d565041b5b555bf

I've resisted bringing in a third-party table library for the moment
because they all seem to have massive feature sets or quirks:

- https://github.com/BurntSushi/tabwriter
- https://github.com/phsym/prettytable-rs
- https://github.com/zhiburt/tabled

Some notes about the current implementation:

- NAME column is padded to 20 chars and will no longer align if it
  overflow but it will still be space delimited
- NAME is already sorted for us and the test uses header/footer entries
  to prevent any future regressions.
- PID column is right align to account variable length values

This was quick and easy to unit test with the helpers that Yannik
recently added. I opted to use real `process-compose` output, rather
than `ProcessStates` fixtures, because it's less fiddly and more
realistic than constructing all of the fields manually.